+++ neuronquant/data/ziplinesources/live/forex.py
@@ -19,6 +19,7 @@ Tools to generate data sources.
 """
 import time
 import datetime
+import pytz
 import pandas as pd
 
 from zipline.gens.utils import hash_args
@@ -84,18 +85,18 @@ class DataLiveSource(DataSource):
         '''
         Only return when we reach given datetime
         '''
-        current_dt = datetime.datetime.now()
-        while (current_dt.minute < dt.minute) or (current_dt.hour < dt.hour) :
+        now = datetime.datetime.now(pytz.utc)
+        while (now.minute < dt.minute) or (now.hour < dt.hour) :
             time.sleep(15)
-            current_dt = datetime.datetime.now()
-            log.info('Waiting {} / {}'.format(current_dt, dt))
+            now = datetime.datetime.now(pytz.utc)
+            log.info('Waiting {} / {}'.format(now, dt))
 
     def _get_updated_index(self):
         '''
         truncate past dates in index
         '''
         late_index = self.data['index']
-        current_dt = datetime.datetime.now()
+        current_dt = datetime.datetime.now(pytz.utc)
         selector = (late_index.day > current_dt.day) \
                 | ((late_index.day == current_dt.day) & (late_index.hour > current_dt.hour)) \
                 | ((late_index.day == current_dt.day) & (late_index.hour == current_dt.hour) & (late_index.minute >= current_dt.minute))