+++ server/db_request.sh
@@ -1,0 +1,1 @@
+http localhost:8080/dashboard/number data==Open table==Quotes field==Ticker value==aapl