+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 23-04-2013
 ----------
 - Errors module
+- Many many fixes
+- New RESTful API along of the broker
+- Better live integration, with team_dashboard integration
 
 10-04-2013
 ----------